requirements.txt
flask
python-dotenv
ldap3
boto3
